Move Validation Functions to a Separate Folder
